Split downgrade view from application instance admin pages
